moved setting panel into own file
